Change memory manager classes to use CRTP
