+++ src/backend/cpu/memory.cpp
@@ -12,8 +12,6 @@
 #include <types.hpp>
 #include <platform.hpp>
 #include <queue.hpp>
-#include <memory>
-#include <MemoryManager.hpp>
 
 #ifndef AF_MEM_DEBUG
 #define AF_MEM_DEBUG 0
@@ -25,88 +23,34 @@
 
 namespace cpu
 {
-
-class MemoryManager  : public common::MemoryManager
-{
-    int getActiveDeviceId();
-    size_t getMaxMemorySize(int id);
-public:
-    MemoryManager();
-    void *nativeAlloc(const size_t bytes);
-    void nativeFree(void *ptr);
-    ~MemoryManager()
-    {
-        common::lock_guard_t lock(this->memory_mutex);
-        for (int n = 0; n < getDeviceCount(); n++) {
-            cpu::setDevice(n);
-            this->garbageCollect();
-        }
-    }
-};
-
-int MemoryManager::getActiveDeviceId()
-{
-    return cpu::getActiveDeviceId();
-}
-
-size_t MemoryManager::getMaxMemorySize(int id)
-{
-    return cpu::getDeviceMemorySize(id);
-}
-
-MemoryManager::MemoryManager() :
-    common::MemoryManager(getDeviceCount(), common::MAX_BUFFERS, AF_MEM_DEBUG || AF_CPU_MEM_DEBUG)
-{
-    this->setMaxMemorySize();
-}
-
-
-void *MemoryManager::nativeAlloc(const size_t bytes)
-{
-    void *ptr = malloc(bytes);
-    if (!ptr) AF_ERROR("Unable to allocate memory", AF_ERR_NO_MEM);
-    return ptr;
-}
-
-void MemoryManager::nativeFree(void *ptr)
-{
-    return free((void *)ptr);
-}
-
-static MemoryManager &getMemoryManager()
-{
-    static MemoryManager instance;
-    return instance;
-}
-
 void setMemStepSize(size_t step_bytes)
 {
-    getMemoryManager().setMemStepSize(step_bytes);
+    memoryManager().setMemStepSize(step_bytes);
 }
 
 size_t getMemStepSize(void)
 {
-    return getMemoryManager().getMemStepSize();
+    return memoryManager().getMemStepSize();
 }
 
 size_t getMaxBytes()
 {
-    return getMemoryManager().getMaxBytes();
+    return memoryManager().getMaxBytes();
 }
 
 unsigned getMaxBuffers()
 {
-    return getMemoryManager().getMaxBuffers();
+    return memoryManager().getMaxBuffers();
 }
 
 void garbageCollect()
 {
-    getMemoryManager().garbageCollect();
+    memoryManager().garbageCollect();
 }
 
 void printMemInfo(const char *msg, const int device)
 {
-    getMemoryManager().printInfo(msg, device);
+    memoryManager().printInfo(msg, device);
 }
 
 template<typename T>
@@ -115,10 +59,10 @@ T* memAlloc(const size_t &elements)
     T *ptr = nullptr;
 
     try {
-        ptr = (T *)getMemoryManager().alloc(elements * sizeof(T), false);
+        ptr = (T *)memoryManager().alloc(elements * sizeof(T), false);
     } catch(...) {
         getQueue().sync();
-        ptr = (T *)getMemoryManager().alloc(elements * sizeof(T), false);
+        ptr = (T *)memoryManager().alloc(elements * sizeof(T), false);
     }
     return ptr;
 }
@@ -128,10 +72,10 @@ void* memAllocUser(const size_t &bytes)
     void *ptr = nullptr;
 
     try {
-        ptr = getMemoryManager().alloc(bytes, true);
+        ptr = memoryManager().alloc(bytes, true);
     } catch(...) {
         getQueue().sync();
-        ptr = getMemoryManager().alloc(bytes, true);
+        ptr = memoryManager().alloc(bytes, true);
     }
     return ptr;
 }
@@ -139,52 +83,52 @@ void* memAllocUser(const size_t &bytes)
 template<typename T>
 void memFree(T *ptr)
 {
-    return getMemoryManager().unlock((void *)ptr, false);
+    return memoryManager().unlock((void *)ptr, false);
 }
 
 void memFreeUser(void *ptr)
 {
-    getMemoryManager().unlock((void *)ptr, true);
+    memoryManager().unlock((void *)ptr, true);
 }
 
 void memLock(const void *ptr)
 {
-    getMemoryManager().userLock((void *)ptr);
+    memoryManager().userLock((void *)ptr);
 }
 
 bool isLocked(const void *ptr)
 {
-    return getMemoryManager().isUserLocked((void *)ptr);
+    return memoryManager().isUserLocked((void *)ptr);
 }
 
 void memUnlock(const void *ptr)
 {
-    getMemoryManager().userUnlock((void *)ptr);
+    memoryManager().userUnlock((void *)ptr);
 }
 
 void deviceMemoryInfo(size_t *alloc_bytes, size_t *alloc_buffers,
                       size_t *lock_bytes,  size_t *lock_buffers)
 {
     getQueue().sync();
-    getMemoryManager().bufferInfo(alloc_bytes, alloc_buffers,
+    memoryManager().bufferInfo(alloc_bytes, alloc_buffers,
                                   lock_bytes,  lock_buffers);
 }
 
 template<typename T>
 T* pinnedAlloc(const size_t &elements)
 {
-    return (T *)getMemoryManager().alloc(elements * sizeof(T), false);
+    return (T *)memoryManager().alloc(elements * sizeof(T), false);
 }
 
 template<typename T>
 void pinnedFree(T* ptr)
 {
-    return getMemoryManager().unlock((void *)ptr, false);
+    return memoryManager().unlock((void *)ptr, false);
 }
 
 bool checkMemoryLimit()
 {
-    return getMemoryManager().checkMemoryLimit();
+    return memoryManager().checkMemoryLimit();
 }
 
 #define INSTANTIATE(T)                                      \
@@ -206,4 +150,45 @@ INSTANTIATE(uintl)
 INSTANTIATE(ushort)
 INSTANTIATE(short )
 
+MemoryManager::MemoryManager()
+    : common::MemoryManager<cpu::MemoryManager>(getDeviceCount(), common::MAX_BUFFERS,
+                                                AF_MEM_DEBUG || AF_CPU_MEM_DEBUG)
+{
+    this->setMaxMemorySize();
+}
+
+MemoryManager::~MemoryManager()
+{
+    common::lock_guard_t lock(this->memory_mutex);
+    for (int n = 0; n < cpu::getDeviceCount(); n++) {
+        try {
+            cpu::setDevice(n);
+            garbageCollect();
+        } catch(AfError err) {
+            continue; // Do not throw any errors while shutting down
+        }
+    }
+}
+
+int MemoryManager::getActiveDeviceId()
+{
+    return cpu::getActiveDeviceId();
+}
+
+size_t MemoryManager::getMaxMemorySize(int id)
+{
+    return cpu::getDeviceMemorySize(id);
+}
+
+void *MemoryManager::nativeAlloc(const size_t bytes)
+{
+    void *ptr = malloc(bytes);
+    if (!ptr) AF_ERROR("Unable to allocate memory", AF_ERR_NO_MEM);
+    return ptr;
+}
+
+void MemoryManager::nativeFree(void *ptr)
+{
+    return free((void *)ptr);
+}
 }
